MemStream: avoid truncating size_t to int
Bug: 417788798

+++ base/include/aemu/base/files/MemStream.h
@@ -27,7 +27,7 @@ class MemStream : public Stream {
 public:
     using Buffer = std::vector<char>;
 
-    MemStream(int reserveSize = 512);
+    MemStream(size_t reserveSize = 512);
     MemStream(Buffer&& data);
 
     MemStream(MemStream&& other) = default;
